+++ bugsnag/src/test/java/com/bugsnag/AppenderTest.java
@@ -21,6 +21,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Pattern;
 
 
 /**
@@ -144,10 +145,15 @@ public class AppenderTest {
         assertEquals("gradleTask", config.appType);
         assertFalse(config.shouldAutoCaptureSessions());
 
-        assertEquals(2, config.filters.length);
-        ArrayList<String> filters = new ArrayList<String>(Arrays.asList(config.filters));
+        assertEquals(8, config.redactedKeys.length);
+        List<Pattern> redactedKeys = new ArrayList<Pattern>(Arrays.asList(config.redactedKeys));
+        assertTrue(containsPattern(redactedKeys, "password"));
+        assertTrue(containsPattern(redactedKeys, "credit_card_number"));
+
+        assertEquals(4, config.filters.length);
+        List<String> filters = new ArrayList<String>(Arrays.asList(config.filters));
         assertTrue(filters.contains("password"));
-        assertTrue(filters.contains("credit_card_number"));
+        assertTrue(filters.contains("secret"));
 
         assertEquals(2, config.ignoreClasses.length);
         ArrayList<String> ignoreClasses
@@ -276,13 +282,13 @@ public class AppenderTest {
     @Test
     public void testFilters() {
 
-        // Add some meta data which should be filtered by key name
+        // Add some meta data which should be redacted by key name
         Bugsnag.addThreadMetaData("myTab", "password", "password value");
         Bugsnag.addThreadMetaData("myTab", "credit_card_number", "card number");
-        Bugsnag.addThreadMetaData("myTab", "mysecret", "not filtered");
+        Bugsnag.addThreadMetaData("myTab", "mysecret", "not redacted");
 
         // Send a log message
-        LOGGER.warn("Exception with filtered meta data", new RuntimeException("test"));
+        LOGGER.warn("Exception with redacted meta data", new RuntimeException("test"));
 
         // Check that a report was sent to Bugsnag
         assertEquals(1, delivery.getNotifications().size());
@@ -291,9 +297,9 @@ public class AppenderTest {
         assertTrue(notification.getEvents().get(0).getMetaData().containsKey("myTab"));
         Map<String, Object> myTab = getMetaDataMap(notification, "myTab");
 
-        assertEquals("[FILTERED]", myTab.get("password"));
-        assertEquals("[FILTERED]", myTab.get("credit_card_number"));
-        assertEquals("not filtered", myTab.get("mysecret"));
+        assertEquals("[REDACTED]", myTab.get("password"));
+        assertEquals("[REDACTED]", myTab.get("credit_card_number"));
+        assertEquals("[REDACTED]", myTab.get("mysecret"));
     }
 
     @Test
@@ -316,7 +322,7 @@ public class AppenderTest {
         });
 
         // Send a log message
-        LOGGER.warn("Exception with filtered meta data", new RuntimeException("test"));
+        LOGGER.warn("Exception with redacted meta data", new RuntimeException("test"));
 
         // Check that a report was sent to Bugsnag
         assertEquals(1, delivery.getNotifications().size());
@@ -421,5 +427,22 @@ public class AppenderTest {
     @SuppressWarnings (value = "unchecked")
     private Map<String, Object> getMetaDataMap(Notification notification, String key) {
         return ((Map<String, Object>) notification.getEvents().get(0).getMetaData().get(key));
+    }
+
+    /**
+    * Checks if the given list of objects contains a pattern with the specified regex and flags.
+    *
+    * @param patterns the list of objects to search through
+    * @param regex the regular expression string to match
+    * @param flags the match flags, a bit mask that may include #CASE_INSENSITIVE, #LITERAL}, etc.
+    * @return {@code true} if a pattern with the specified regex and flags is found in the list, {@code false} otherwise
+    */
+    private static boolean containsPattern(List<Pattern> patterns, String regex) {
+        for (Pattern pattern : patterns) {
+            if (pattern.pattern().equals(regex)) {
+                return true;
+            }
+        }
+        return false;
     }
 }