+++ bugsnag/src/test/java/com/bugsnag/JavaxServletCallbackTest.java
@@ -100,10 +100,10 @@ public class JavaxServletCallbackTest {
         assertEquals("some-data-1,some-data-2", headers.get("X-Custom-Header"));
 
         // Make sure that actual Authorization header value is not in the report
-        assertEquals("[FILTERED]", headers.get("Authorization"));
+        assertEquals("[REDACTED]", headers.get("Authorization"));
 
         // Make sure that actual cookies are not in the report
-        assertEquals("[FILTERED]", headers.get("Cookie"));
+        assertEquals("[REDACTED]", headers.get("Cookie"));
 
         assertTrue(request.containsKey("params"));
         Map<String, String[]> params = (Map<String, String[]>) request.get("params");