+++ features/fixtures/scenarios/src/main/java/com/bugsnag/mazerunner/scenarios/ManualFilterScenario.java
@@ -5,7 +5,7 @@ import com.bugsnag.Report;
 import com.bugsnag.callbacks.Callback;
 
 /**
- * Sends a handled exception to Bugsnag, which contains metadata that should be filtered
+ * Sends a handled exception to Bugsnag, which contains metadata that should be redacted
  */
 public class ManualFilterScenario extends Scenario {
 
@@ -15,14 +15,13 @@ public class ManualFilterScenario extends Scenario {
 
     @Override
     public void run() {
-
         bugsnag.setFilters("foo");
 
         bugsnag.notify(generateException(), new Callback() {
             @Override
             public void beforeNotify(Report report) {
-                report.addToTab("user", "foo", "hunter2");
                 report.addToTab("custom", "foo", "hunter2");
+                report.addToTab("user", "abcd", "hunter2");
                 report.addToTab("custom", "bar", "hunter2");
             }
         });