+++ features/fixtures/scenarios/src/main/java/com/bugsnag/mazerunner/scenarios/ManualRedactedKeysScenario.java
@@ -3,24 +3,28 @@ package com.bugsnag.mazerunner.scenarios;
 import com.bugsnag.Bugsnag;
 import com.bugsnag.Report;
 import com.bugsnag.callbacks.Callback;
+import java.util.regex.Pattern;
 
 /**
  * Sends a handled exception to Bugsnag, which contains metadata that should be redacted
  */
-public class AutoFilterScenario extends Scenario {
+public class ManualRedactedKeysScenario extends Scenario {
 
-    public AutoFilterScenario(Bugsnag bugsnag) {
+    public ManualRedactedKeysScenario(Bugsnag bugsnag) {
         super(bugsnag);
     }
 
     @Override
     public void run() {
+        Pattern pattern = Pattern.compile("[a-zA-Z]{4}");
+        bugsnag.setRedactedKeys(Pattern.compile("foo"), pattern);
+
         bugsnag.notify(generateException(), new Callback() {
             @Override
             public void beforeNotify(Report report) {
-                report.addToTab("user", "password", "hunter2");
-                report.addToTab("custom", "password", "hunter2");
                 report.addToTab("custom", "foo", "hunter2");
+                report.addToTab("user", "abcd", "hunter2");
+                report.addToTab("custom", "bar", "hunter2");
             }
         });
     }